// src/app/models/planet.ts
export interface Planet {
  climate: string;
  created: string;
  diameter: string;
  edited: string;
  films: string;
  gravity: string;
  name: string;
  orbital_period: string;
  population: string;
  residents: string;
  rotation_period: string;
  starships: string;
  surface_water: string;
  url: string;
  terrain: string;
}